http.Pusher only exists in go1.8+
